refactor(types): move WorkingMemory wire types to origin-types
