+++ crates/origin-types/Cargo.toml
@@ -13,5 +13,6 @@ categories = ["data-structures", "api-bindings"]
 
 [dependencies]
 anyhow = { workspace = true }
+chrono = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
