+++ crates/origin-types/src/lib.rs
@@ -14,6 +14,7 @@ pub mod requests;
 pub mod responses;
 pub mod sources;
 pub mod system_info;
+pub mod working_memory;
 
 // Re-export commonly used types at crate root for convenience.
 pub use entities::{
